requirements.txt
pyyaml==6.0.2
rich==13.9.4
